wip - refactoring out hacker-card
